Renamed 'BubbleSection' to 'Section'
