+++ pages/index.vue
@@ -15,7 +15,7 @@
     <div class="hero is-primary is-bold">
       <div class="hero-body">
         <div class="container">
-          <BubbleSection
+          <Section
             v-for="(category, n) in everyCategory"
             :key="n"
             v-bind="category"
@@ -27,7 +27,7 @@
     <div class="hero is-light is-bold">
       <div class="hero-body">
         <div class="container">
-          <BubbleSection
+          <Section
             v-for="(bot, n) in everyBot"
             :key="n"
             v-bind="bot"
@@ -39,7 +39,7 @@
     <div class="hero is-dark is-bold">
       <div class="hero-body">
         <div class="container">
-          <BubbleSection
+          <Section
             v-for="(project, n) in everyProject"
             :key="n"
             v-bind="project"
